fix pk ItemPedido.java
There was an issue were the ItemPedido primary key was the Produto primary key, which was a issue because this value is expected to be repeat as we can make lots of different orders of the same product.

To fix that, was added a new variable Long id, which will be the new primary key, due to the high complexity to implement (codProduto, idOrcamento) tuple to be the primary key.